refactor(XRB) adds utils and pow folders
