+++ src/XRB/index.js
@@ -1,4 +1,3 @@
-import BigNumber from 'bignumber.js';
 import { blake2b, blake2bFinal, blake2bInit, blake2bUpdate } from 'blakejs';
 import nacl from '../../vendor/nacl';
 import {
@@ -13,44 +12,9 @@ import {
   hexToUint8,
   intToUint8,
 } from '../utils/uint';
+import { isValidAccount, isValidHash, subtract, rawToHex } from './utils';
 import { convertToRaw } from '../utils/getConversion';
 import { areArraysEqual, arrayCrop, arrayExtend } from '../utils/arrays';
-
-export const isValidHash = (hash, bytes = 32) => {
-  if (bytes === 16) return /[0-9A-F]{32}\b/i.test(hash) && hash.length === (bytes * 2);
-  if (bytes === 32) return /[0-9A-F]{64}\b/i.test(hash) && hash.length === (bytes * 2);
-  if (bytes === 64) return /[0-9A-F]{128}/i.test(hash) && hash.length === (bytes * 2);
-  throw new Error(`Bytes must be 16, 32 or 64, ${bytes} supplied`);
-};
-
-export const isValidAccount = account =>
-  (account.startsWith('xrb_1') || account.startsWith('xrb_3')) && account.length === 64;
-
-
-// Use for RAW
-const subtract = (base, minus) => {
-  let value = new BigNumber(base.toString());
-  const bigMinus = new BigNumber(minus.toString());
-  if (bigMinus.greaterThan(value)) {
-    throw new Error('Subtraction will result in negative value');
-  }
-
-  value = value.minus(bigMinus);
-  value = value.toFixed(0);
-  return value;
-};
-
-// Use for RAW
-const rawToHex = (raw) => {
-  let value = new BigNumber(raw.toString());
-  value = value.toString(16).toUpperCase();
-  if (value.length < 32) {
-    for (let n = value.length; n < 32; n += 1) {
-      value = `0${value}`;
-    }
-  }
-  return value;
-};
 
 export const getAccountKey = (account) => {
   if (!isValidAccount(account)) {
@@ -168,107 +132,6 @@ export const seedKeys = (seedHex, count = 1) => {
     keys.push(uint8ToHex(blake2bFinal(context)));
   }
   return keys;
-};
-
-const powInitiate = (_threads, workerPath = '') => {
-  if (!window || !window.Worker || !window.navigator) return false;
-  let threads = _threads;
-  if (Number.isNaN(Number(threads))) {
-    threads = window.navigator.hardwareConcurrency - 1;
-  }
-  const workers = [];
-  for (let i = 0; i < threads; i += 1) {
-    workers[i] = new Worker(`${workerPath}rai.pow.js`);
-  }
-  return workers;
-};
-
-const powThreshold = (Uint8Array) => {
-  if (
-    (Uint8Array[0] === 255)
-    && (Uint8Array[1] === 255)
-    && (Uint8Array[2] === 255)
-    && (Uint8Array[3] >= 192)
-  ) {
-    return true;
-  }
-  return false;
-};
-
-const powValidate = (powHex, hashHex) => {
-  if (!isValidHash(hashHex)) {
-    throw new Error('Invalid: hash hex is not a valid hash');
-  }
-
-  const hash = hexToUint8(hashHex);
-  const isValidPOW = /^[0123456789ABCDEFabcdef]+$/.test(powHex);
-  if (isValidPOW && (powHex.length === 16)) {
-    const pow = hexToUint8(powHex);
-    const context = blake2bInit(8, null);
-    blake2bUpdate(context, pow.reverse());
-    blake2bUpdate(context, hash);
-    const check = blake2bFinal(context).reverse();
-
-    if (powThreshold(check)) {
-      return true;
-    }
-
-    return false;
-  }
-
-  throw new Error('Invalid work');
-};
-
-const powTerminate = (workers) => {
-  const threads = workers.length;
-  for (let i = 0; i < threads; i += 1) {
-    workers[i].terminate();
-  }
-};
-
-const onWorkerMessage = (worker, workers, hash, callback) => (e) => {
-  const result = e.data;
-  if (result) {
-    powTerminate(workers);
-    callback(result);
-  } else worker.postMessage(hash);
-};
-
-const powCallback = (workers, hash, callback) => {
-  if ((hash instanceof Uint8Array) && (hash.length === 32) && (typeof callback === 'function')) {
-    const threads = workers.length;
-    for (let i = 0; i < threads; i += 1) {
-      const worker = workers[i];
-      worker.onmessage = onWorkerMessage(worker, workers, hash, callback);
-    }
-  } else if (typeof callback !== 'function') {
-    throw new Error('Invalid callback function');
-  } else {
-    throw new Error('Invalid hash array');
-  }
-};
-
-const powStart = (workers, hash) => {
-  if ((hash instanceof Uint8Array) && (hash.length === 32)) {
-    const threads = workers.length;
-    for (let i = 0; i < threads; i += 1) {
-      workers[i].postMessage(hash);
-    }
-  } else {
-    throw new Error('Invalid hash array');
-  }
-};
-
-const pow = (hashHex, threads, callback, workerPath) => {
-  if (!isValidHash(hashHex)) {
-    throw new Error('Invalid hash');
-  }
-
-  const hash = hexToUint8(hashHex);
-  const workers = powInitiate(threads, workerPath);
-  powStart(workers, hash);
-
-  return powCallback(workers, hash, callback);
 };
 
 /**
@@ -457,8 +320,6 @@ export default {
   change,
   checkSignature,
   open,
-  pow,
-  powValidate,
   receive,
   seedKey,
   seedKeys,
